refactor(runtime): Re-home FCP queue SPI ports
Move the FCP-backed queue download, insert, and completion SPI implementations from runtime.admin to runtime.endpoints.fcp.

Update the admin factory wiring, endpoint package docs, moved package-local tests, and LegacyRuntimePorts ownership assertions without changing queue behavior or runtime-spi.
