+++ src/main/java/network/crypta/runtime/endpoints/fcp/LegacyQueueDownloadPort.java
@@ -1,4 +1,4 @@
-package network.crypta.runtime.admin;
+package network.crypta.runtime.endpoints.fcp;
 
 import java.io.IOException;
 import java.util.Objects;
