+++ src/test/java/network/crypta/runtime/endpoints/fcp/LegacyQueueInsertPortTest.java
@@ -1,4 +1,4 @@
-package network.crypta.runtime.admin;
+package network.crypta.runtime.endpoints.fcp;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
